// src/pages/detalhes/interfaces.ts
export interface IResult {

    changePercent24Hr: string
    explorer: string
    id: string
    marketCapUsd: string
    maxSupply: string
    name: string
    priceUsd: string
    rank: string
    supply: string
    symbol: string
    volumeUsd24Hr: string
    vwap24Hr: string

}

export interface IMoeda {

    logo?: string
    nome?: string
    valorMercado?: string
    preco?: string
    volume?: string
    oscilacao?: string

}